// api/src/types/interface.d.ts
type Faculty = {
    facultyCode?: string;
    facultyNameKh?: string;
    facultyNameEn?: string;
    changedByUser?: string;
    clientIp?: string;
    sessionInfo?: string;
    status?: 'Active' | 'Inactive';
}
// ...
export interface User {
    id?: number;
    username?: string;
    email: string;
    password: string;
    isActive?: boolean | true;
    assignType?: 'Faculty' | 'Student';
    assignTo?: number;
}
export interface SearchParams{
    keyword?: string;
    page?: number;
    limit?: number;
} 
export interface Student{
    studentCode?: string;
    studentNameKh?: string;
    studentNameEn?: string;
    dob?: Date;
    gender?: 'Male' | 'Female';
    email?: string;
    phoneNumber?: string;
    password?: string;
    changedByUser?: string;
    clientIp?: string;
    sessionInfo?: string;
    status?: 'Active' | 'Inactive';
}
// ...
export interface UserSignUp {
    username: string;
    email: string;
    password: string;
    isActive?: boolean;
    assignType?: string;
    assignTo?: number;
}